Enda mer refaktorering
Enda mer refaktorering
